requirements.txt
#fastapi
#uvicorn[standard]
#python-dotenv
#pydantic
#langchain
#chromadb
#sentence-transformers
#openai
#groq
#pdfplumber
#pytesseract
#pillow
#requests
#streamlit
#sqlalchemy
#psycopg2-binary
#python-multipart
#yfinance

#yalog


# Web Framework & Serialization
fastapi
uvicorn[standard]
pydantic
python-dotenv

# AI & Tooling
google-genai
langchain-core  # For Agent orchestration (optional, but good practice)
requests
pandas
pytest
pytest-mock
ruff
chromadb
weaviate-client
scikit-learn

# Groq client for Groq LLM provider
groq

# Document Handling (Lite version)
# pypdf (for doc_analyzer.py)
# pillow (for ocr_tool.py)